changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11790 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Histogram Objects/PlotFFT/ORPlotFFT.h
@@ -49,7 +49,7 @@
 - (void) setShowImaginary:(BOOL)aFlag;
 - (BOOL) showPowerSpectrum;
 - (void) setShowPowerSpectrum:(BOOL)aFlag;
-- (unsigned) numberChans;
+- (NSUInteger) numberChans;
 
 #pragma mark ¥¥¥Data Management
 - (void) clear;
